Add lower function in SQL select
There was a conflict between uppercase and lowercase letters when performing queries. We added the LOWER() function so that, at query time, the submitted arguments and database data are converted to lowercase so that the comparison is true.

Example:
where xyz = 'XyZ' - This is false
where LOWER(Xyz) = 'xYz' - This is true